requirements.txt
django==4.1.4
django-environ==0.9.0
psycopg2-binary==2.9.5
requests==2.28.1 
django-herald==0.3.0